requirements.txt
psycopg2-binary
Django==4.0.5
zappa==0.54.1
django-storages==1.12.3
django-model-utils==4.2.0
pydash==5.1.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
dj-rest-auth==2.2.4
django-allauth==0.50.0
drf-spectacular==0.22.1
django-extensions==3.1.5
django-environ==0.8.1
scrypt==0.8.20
Pillow==9.2.0
django-filter==22.1
django-rest-swagger
djangorestframework-simplejwt